+++ sdk/resourcemanager/desktopvirtualization/armdesktopvirtualization/applicationgroups_client.go
@@ -48,7 +48,7 @@ func NewApplicationGroupsClient(subscriptionID string, credential azcore.TokenCr
 // CreateOrUpdate - Create or update an applicationGroup.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-09-09
+// Generated from API version 2023-01-30-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - applicationGroupName - The name of the application group
 //   - applicationGroup - Object containing ApplicationGroup definitions.
@@ -89,7 +89,7 @@ func (client *ApplicationGroupsClient) createOrUpdateCreateRequest(ctx context.C
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-09")
+	reqQP.Set("api-version", "2023-01-30-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, applicationGroup)
@@ -107,7 +107,7 @@ func (client *ApplicationGroupsClient) createOrUpdateHandleResponse(resp *http.R
 // Delete - Remove an applicationGroup.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-09-09
+// Generated from API version 2023-01-30-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - applicationGroupName - The name of the application group
 //   - options - ApplicationGroupsClientDeleteOptions contains the optional parameters for the ApplicationGroupsClient.Delete
@@ -147,7 +147,7 @@ func (client *ApplicationGroupsClient) deleteCreateRequest(ctx context.Context, 
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-09")
+	reqQP.Set("api-version", "2023-01-30-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -156,7 +156,7 @@ func (client *ApplicationGroupsClient) deleteCreateRequest(ctx context.Context, 
 // Get - Get an application group.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-09-09
+// Generated from API version 2023-01-30-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - applicationGroupName - The name of the application group
 //   - options - ApplicationGroupsClientGetOptions contains the optional parameters for the ApplicationGroupsClient.Get method.
@@ -195,7 +195,7 @@ func (client *ApplicationGroupsClient) getCreateRequest(ctx context.Context, res
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-09")
+	reqQP.Set("api-version", "2023-01-30-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -212,7 +212,7 @@ func (client *ApplicationGroupsClient) getHandleResponse(resp *http.Response) (A
 
 // NewListByResourceGroupPager - List applicationGroups.
 //
-// Generated from API version 2022-09-09
+// Generated from API version 2023-01-30-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - options - ApplicationGroupsClientListByResourceGroupOptions contains the optional parameters for the ApplicationGroupsClient.NewListByResourceGroupPager
 //     method.
@@ -260,7 +260,7 @@ func (client *ApplicationGroupsClient) listByResourceGroupCreateRequest(ctx cont
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-09")
+	reqQP.Set("api-version", "2023-01-30-preview")
 	if options != nil && options.Filter != nil {
 		reqQP.Set("$filter", *options.Filter)
 	}
@@ -289,7 +289,7 @@ func (client *ApplicationGroupsClient) listByResourceGroupHandleResponse(resp *h
 
 // NewListBySubscriptionPager - List applicationGroups in subscription.
 //
-// Generated from API version 2022-09-09
+// Generated from API version 2023-01-30-preview
 //   - options - ApplicationGroupsClientListBySubscriptionOptions contains the optional parameters for the ApplicationGroupsClient.NewListBySubscriptionPager
 //     method.
 func (client *ApplicationGroupsClient) NewListBySubscriptionPager(options *ApplicationGroupsClientListBySubscriptionOptions) *runtime.Pager[ApplicationGroupsClientListBySubscriptionResponse] {
@@ -332,7 +332,7 @@ func (client *ApplicationGroupsClient) listBySubscriptionCreateRequest(ctx conte
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-09")
+	reqQP.Set("api-version", "2023-01-30-preview")
 	if options != nil && options.Filter != nil {
 		reqQP.Set("$filter", *options.Filter)
 	}
@@ -353,7 +353,7 @@ func (client *ApplicationGroupsClient) listBySubscriptionHandleResponse(resp *ht
 // Update - Update an applicationGroup.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-09-09
+// Generated from API version 2023-01-30-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - applicationGroupName - The name of the application group
 //   - options - ApplicationGroupsClientUpdateOptions contains the optional parameters for the ApplicationGroupsClient.Update
@@ -393,7 +393,7 @@ func (client *ApplicationGroupsClient) updateCreateRequest(ctx context.Context, 
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-09")
+	reqQP.Set("api-version", "2023-01-30-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if options != nil && options.ApplicationGroup != nil {