+++ sdk/resourcemanager/desktopvirtualization/armdesktopvirtualization/msiximages_client.go
@@ -46,7 +46,7 @@ func NewMsixImagesClient(subscriptionID string, credential azcore.TokenCredentia
 
 // NewExpandPager - Expands and Lists MSIX packages in an Image, given the Image Path.
 //
-// Generated from API version 2022-09-09
+// Generated from API version 2023-01-30-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - hostPoolName - The name of the host pool within the specified resource group
 //   - msixImageURI - Object containing URI to MSIX Image
@@ -99,7 +99,7 @@ func (client *MsixImagesClient) expandCreateRequest(ctx context.Context, resourc
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-09-09")
+	reqQP.Set("api-version", "2023-01-30-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, msixImageURI)